[android] fix after rebase
